fix: Fix issue with refactoring of backend classes. SCRUM-337 #test 20m
