+++ backend/src/main/java/stud/ntnu/backend/repository/EmailTokenRepository.java
@@ -2,7 +2,7 @@ package stud.ntnu.backend.repository;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
-import stud.ntnu.backend.model.EmailToken;
+import stud.ntnu.backend.model.user.EmailToken;
 
 import java.util.Optional;
