+++ backend/src/main/java/stud/ntnu/backend/service/AuthService.java
@@ -10,12 +10,12 @@ import org.springframework.security.core.context.SecurityContextHolder;
 import org.springframework.security.crypto.password.PasswordEncoder;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
-import stud.ntnu.backend.dto.AuthRequestDto;
-import stud.ntnu.backend.dto.AuthResponseDto;
-import stud.ntnu.backend.dto.RegisterRequestDto;
-import stud.ntnu.backend.model.EmailToken;
-import stud.ntnu.backend.model.Role;
-import stud.ntnu.backend.model.User;
+import stud.ntnu.backend.dto.auth.AuthRequestDto;
+import stud.ntnu.backend.dto.auth.AuthResponseDto;
+import stud.ntnu.backend.dto.auth.RegisterRequestDto;
+import stud.ntnu.backend.model.user.EmailToken;
+import stud.ntnu.backend.model.user.Role;
+import stud.ntnu.backend.model.user.User;
 import stud.ntnu.backend.repository.EmailTokenRepository;
 import stud.ntnu.backend.repository.UserRepository;
 import stud.ntnu.backend.util.JwtUtil;
